Final API for Subproduct add and create

+++ src/main/java/com/dbtest/dbtest/Entity/Subproduct.java
@@ -28,10 +28,10 @@ public class Subproduct {
     private VendorEntity vendorDetails;
 
     @Column(name="productsellingprice")
-    private float productSellingPrice;
+    private Double productSellingPrice;
 
     @Column(name="productpurchaseprice")
-    private float productPurchasePrice;
+    private Double productPurchasePrice;
 
     @Column(name="lastupdatedby")
     private String lastUpdatedBy;
